requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor-5==0.2.13
django-jazzmin==3.0.0
django-modeltranslation==0.19.3
numpy==1.26.4
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
